requirements.txt.txt
streamlit
transformer
torch